refactor: rename NavigationDrawer to ShortcutDrawer
