+++ src/components/ShortcutDrawer/Icons.android.tsx
