+++ src/screens/SearchScreen.tsx
@@ -4,7 +4,7 @@ import { Container, Header, Input, Left, Form, Item, Icon } from 'native-base'
 import { DrawerActions } from '@react-navigation/native'
 
 import SearchResultList from './SearchResultList'
-import NavigationDrawer from '../components/NavigationDrawer'
+import ShortcutDrawer from '../components/ShortcutDrawer'
 
 type SearchScreenProps = {
   navigation: any;
@@ -47,7 +47,7 @@ const SearchScreen: FunctionComponent<SearchScreenProps> = ( { navigation } ) =>
           setInitPageCount={setPageCount}
         />
       </View>
-      <NavigationDrawer />
+      <ShortcutDrawer />
     </Container>
   )
 }
